refactor : 레이아웃, 장바구니 페이지 개선
