+++ src/pages/Cart.jsx
@@ -1,31 +1,26 @@
 import React from "react";
-import { useRecoilState, useResetRecoilState } from "recoil";
+import { useRecoilValue, useResetRecoilState } from "recoil";
 import { Link } from "react-router-dom";
 import styled from "styled-components";
 import { useNavigate } from "react-router-dom";
 import { cartItemsState } from "../recoil";
-import { checkImageUrl } from "../utils";
 
 import { Button } from "../components/common";
 import Layout from "../layout/Layout";
 
 import cartNullIcon from "../assets/cart_null_icon.svg";
-import iconClose from "../assets/icon_close.svg";
+
+import CartList from "../components/CartList/CartList";
+import { priceDivide } from "../utils";
 
 export default function Cart() {
-  const [cartItem, setCartItem] = useRecoilState(cartItemsState);
+  const cartItem = useRecoilValue(cartItemsState);
   const navigate = useNavigate();
 
   const handleButtonClick = () => {
     if (cartItem.length !== 0) {
       navigate("/chat");
-      console.log();
     }
-  };
-
-  const removeItem = (itemId) => {
-    const updatedItems = cartItem.filter((item) => item.id !== itemId);
-    setCartItem(updatedItems);
   };
 
   const calculateTotalCount = () => {
@@ -42,10 +37,6 @@ export default function Cart() {
     cartRest();
   };
 
-  const priceDivide = (price) => {
-    return price.toString().replace(/\B(?<!\.\d*)(?=(\d{3})+(?!\d))/g, ",");
-  };
-
   return (
     <Layout reduceTop="true">
       <h2 className="a11y-hidden">장바구니</h2>
@@ -60,27 +51,7 @@ export default function Cart() {
           ) : (
             <ul>
               {cartItem.map((item) => (
-                <CartProductItem key={item.id}>
-                  <CartUserInfo>
-                    <img src={checkImageUrl(item.userImage, "profile")} alt="판매자 프로필 이미지" />
-                    <strong>{item.userName}</strong>
-                    <button onClick={() => removeItem(item.id)}>
-                      <img src={iconClose} alt="상품 삭제 버튼" />
-                    </button>
-                  </CartUserInfo>
-                  <CartProductInfo>
-                    <img src={checkImageUrl(item.productImage, "post")} alt="" />
-                    <CartProductDesc>
-                      <p>No. {item.id}</p>
-                      <strong>{item.productName}</strong>
-                      <p>{priceDivide(item.productPrice)} 원</p>
-                      <CartProductTotal>
-                        <strong>{priceDivide(item.productPrice * item.productCount)} 원</strong>
-                        <span>수량 {item.productCount}개</span>
-                      </CartProductTotal>
-                    </CartProductDesc>
-                  </CartProductInfo>
-                </CartProductItem>
+                <CartList item={item} key={item.id} />
               ))}
             </ul>
           )}
@@ -127,106 +98,6 @@ const CartWrap = styled.section`
 const CartLeft = styled.section`
   width: 60%;
   min-height: 550px;
-`;
-
-const CartProductItem = styled.li`
-  border: 1px solid #ebebeb;
-  border-radius: 0 0 4px 4px;
-  margin-bottom: 28px;
-
-  &:last-child {
-    margin-bottom: 0;
-  }
-`;
-
-const CartUserInfo = styled.div`
-  width: 100%;
-  height: 50px;
-  background-color: var(--gray100-color);
-  display: flex;
-  align-items: center;
-  padding: 10px 16px;
-  box-sizing: border-box;
-
-  & > img {
-    width: 30px;
-    height: 30px;
-    border-radius: 50%;
-    margin-right: 10px;
-  }
-
-  & > strong {
-    font-size: 16px;
-    font-family: var(--font--semibold);
-  }
-
-  & > button {
-    width: 40px;
-    height: 40px;
-    margin-left: auto;
-    cursor: pointer;
-
-    & > img {
-      display: block;
-      margin: 0 auto;
-    }
-  }
-`;
-
-const CartProductInfo = styled.div`
-  width: 100%;
-  padding: 18px 24px 20px;
-  box-sizing: border-box;
-  display: flex;
-  gap: 28px;
-
-  & > img {
-    width: 150px;
-    aspect-ratio: 1 / 1;
-    object-fit: cover;
-  }
-`;
-
-const CartProductDesc = styled.div`
-  width: 100%;
-  p {
-    margin-bottom: 14px;
-
-    &:first-child {
-      margin-top: 5px;
-      font-size: 14px;
-      color: var(--gray400-color);
-    }
-    &:last-child {
-      font-size: 16px;
-    }
-  }
-
-  strong {
-    display: block;
-    font-family: var(--font--semibold);
-    font-size: 20px;
-    margin-bottom: 14px;
-  }
-`;
-
-const CartProductTotal = styled.div`
-  width: 100%;
-  border-top: 1px solid var(--gray100-color);
-  padding-top: 18px;
-  display: flex;
-  justify-content: space-between;
-
-  strong {
-    font-family: var(--font--Bold);
-    font-size: 20px;
-  }
-
-  span {
-    font-family: var(--font--Medium);
-    color: var(--dark-sub-color);
-    font-size: 18px;
-  }
 `;
 
 const CartNull = styled.div`
