+++ src/utils/index.js
@@ -1,5 +1,6 @@
 import checkImageUrl from "./checkImageUrl";
 import { setLoginCookie, getLoginCookie, removeLoginCookie } from "./loginCookie";
 import BASE_URL from "./BASE_URL";
+import priceDivide from "./priceDivide";
 
-export { checkImageUrl, setLoginCookie, getLoginCookie, removeLoginCookie, BASE_URL };
+export { checkImageUrl, setLoginCookie, getLoginCookie, removeLoginCookie, BASE_URL, priceDivide };
